requirements.txt


paho-mqtt
pandas
flask
pytest



